updated code to most recent changes+alex arm added
