+++ src/main/java/frc/robot/subsystems/RobotArm.java
@@ -9,24 +9,16 @@ import frc.robot.subsystems.abstractSubsystems.RobotArmBase;
 // i know that this might break things but i cant help but have a little fun with abstraction as it also allows for easier tweaking of arm subsystems at competion without breaking code
 public class RobotArm extends RobotArmBase {
   /** Creates a new robotArm. */
-  private double avgDeltaVelocity;
-  private double prevVelocity;
+  
   /*private double upperAngle;
   private double lowerAngle;*/
   
   
   @Override
   protected void otherConfigs() {
-    avgDeltaVelocity = 0;
-    prevVelocity = 0;   
+      
   }
     
-  
-
-  @Override
-  public void periodic() {
-      //System.out.println("arm encoder position" + armMotor.getEncoder().getPosition());
-  }
   @Override
   protected void raiseArm() {
     
@@ -37,33 +29,7 @@ public class RobotArm extends RobotArmBase {
       else{
         armMotor.set(0.05);
       }
-    /*
-    // sorry alex your code doesnt work
-    if (!armStatus) {
-      System.out.println("robot arm raise running");
-      armMotor.set(.2);
-
-      double currentVelocity = armMotor.getEncoder().getVelocity();
-      double currentDeltaVelocity = Math.abs(prevVelocity - currentVelocity);
-
-      // update EMA
-      avgDeltaVelocity = avgDeltaVelocity + (DriveConstants.ROBOT_ARM_EXPONENT_WEIGHT * (currentDeltaVelocity - avgDeltaVelocity));
-      
-      if (Math.abs(avgDeltaVelocity - currentDeltaVelocity) > (avgDeltaVelocity / DriveConstants.ROBOT_ARM_DELTA_SENSITIVITY)) {
-        System.out.println("robot arm raise stopping");
-        armMotor.set(0);
-        armStatus = true;
-      }
-
-      prevVelocity = currentVelocity;
-      
-    } else {
-      System.out.println("robot arm already raised");
-      armMotor.set(0);
-      prevVelocity = 0;
-      avgDeltaVelocity = 0;
-    }
-    */
+    
   }
   @Override
   protected void lowerArm()  {
@@ -77,34 +43,25 @@ public class RobotArm extends RobotArmBase {
       0.05
       );
     }
-    /*
-    if (armStatus) { 
-      System.out.println("robot arm lower running");
+    
+  }
 
-      armMotor.set(-.2);
-
-      double currentVelocity = armMotor.getEncoder().getVelocity();
-      double currentDeltaVelocity = Math.abs(prevVelocity - currentVelocity);
-      // update EMA
-      avgDeltaVelocity = avgDeltaVelocity + (DriveConstants.ROBOT_ARM_EXPONENT_WEIGHT * (currentDeltaVelocity - avgDeltaVelocity));
-      
-      if (Math.abs(avgDeltaVelocity - currentDeltaVelocity) > (avgDeltaVelocity / DriveConstants.ROBOT_ARM_DELTA_SENSITIVITY)) {
-        System.out.println("robot arm lower stopped");
-        armMotor.set(0);
-        
-        armStatus = false;
+  @Override
+  public void moveArm(boolean ArmUp) {
+    if (ArmUp) {
+      raiseArm();
+      if(!armStatus){
+        armTimer.reset();
       }
-
-      prevVelocity = currentVelocity;
-      
-
+      armStatus = true;
     } else {
-      System.out.println("robot arm already lowered");
-      armMotor.set(0);
-      prevVelocity = 0;
-      avgDeltaVelocity = 0;
+      lowerArm();
+      if(armStatus){
+        armTimer.reset();
+      }
+      armStatus = false;
     }
-    */
+    
   }
